server/requirements.txt
astroid==2.15.5
beautifulsoup4==4.12.2
black==23.3.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dill==0.3.6
Flask==2.3.2
idna==3.4
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
lyricsgenius==3.0.1
MarkupSafe==2.1.3
mccabe==0.7.0
mypy==1.3.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.6.0
pylint==2.17.4
python-datamuse==2.0.1
python-dotenv==1.0.0
requests==2.31.0
soupsieve==2.4.1
tomli==2.0.1
tomlkit==0.11.8
typing_extensions==4.6.3
urllib3==2.0.3
Werkzeug==2.3.6
wrapt==1.15.0
